Separate Phone and Desktop inventory layouts into different components
