+++ src/app/inventory/Stores.tsx
@@ -1,23 +1,11 @@
-import { scrollToPosition } from 'app/dim-ui/scroll';
-import { t } from 'app/i18next-t';
-import InventoryCollapsibleTitle from 'app/inventory/InventoryCollapsibleTitle';
-import StoreStats from 'app/store-stats/StoreStats';
 import { RootState } from 'app/store/types';
-import clsx from 'clsx';
-import React, { useEffect, useRef, useState } from 'react';
-import Hammer from 'react-hammerjs';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
-import { Frame, Track, View, ViewPager } from 'react-view-pager';
-import StoreHeading from '../character-tile/StoreHeading';
-import ScrollClassDiv from '../dim-ui/ScrollClassDiv';
-import { hideItemPopup } from '../item-popup/item-popup';
-import { storeBackgroundColor } from '../shell/filters';
-import D1ReputationSection from './D1ReputationSection';
-import { InventoryBucket, InventoryBuckets } from './inventory-buckets';
+import DesktopStores from './DesktopStores';
+import { InventoryBuckets } from './inventory-buckets';
+import PhoneStores from './PhoneStores';
 import { bucketsSelector, sortedStoresSelector } from './selectors';
 import { DimStore } from './store-types';
-import { StoreBuckets } from './StoreBuckets';
-import { findItemsByBucket, getCurrentStore, getStore, getVault } from './stores-helpers';
 import './Stores.scss';
 
 interface StoreProps {
@@ -40,13 +28,6 @@ type Props = StoreProps;
  * Display inventory and character headers for all characters and the vault.
  */
 function Stores(this: void, { stores, buckets, isPhonePortrait }: Props) {
-  const vault = getVault(stores)!;
-  const currentStore = getCurrentStore(stores)!;
-
-  const [selectedStoreId, setSelectedStoreId] = useState(currentStore?.id);
-  const [selectedCategoryId, setSelectedCategoryId] = useState<string>('Weapons');
-  const detachedLoadoutMenu = useRef<HTMLDivElement>(null);
-
   useEffect(() => {
     setTimeout(() => {
       /* Set a CSS variable so we can style things based on the height of the header */
@@ -63,247 +44,11 @@ function Stores(this: void, { stores, buckets, isPhonePortrait }: Props) {
     return null;
   }
 
-  const selectedStore = selectedStoreId ? getStore(stores, selectedStoreId)! : currentStore;
-
-  const onViewChange = (indices: number[]) => {
-    setSelectedStoreId(stores[indices[0]].id);
-    hideItemPopup();
-  };
-
-  const handleSwipe: HammerListener = (e) => {
-    const selectedStoreIndex = selectedStoreId
-      ? stores.findIndex((s) => s.id === selectedStoreId)
-      : stores.findIndex((s) => s.current);
-
-    if (e.direction === 2 && selectedStoreIndex < stores.length - 1) {
-      setSelectedStoreId(stores[selectedStoreIndex + 1].id);
-    } else if (e.direction === 4 && selectedStoreIndex > 0) {
-      setSelectedStoreId(stores[selectedStoreIndex - 1].id);
-    }
-  };
-
-  const handleCategoryChange = (category: string) => {
-    if (category === selectedCategoryId) {
-      // If user selects the category they are already on, scroll to top
-      scrollToPosition({ top: 0 });
-      return;
-    }
-    setSelectedCategoryId(category);
-  };
-
-  if (isPhonePortrait) {
-    return (
-      <div
-        className={`inventory-content phone-portrait destiny${selectedStore.destinyVersion}`}
-        role="main"
-        aria-label={t('Header.Inventory')}
-      >
-        <ScrollClassDiv
-          className="store-row store-header"
-          scrollClass="sticky"
-          style={storeBackgroundColor(selectedStore, 0, true, isPhonePortrait)}
-          onTouchStart={(e) => e.stopPropagation()}
-        >
-          <ViewPager>
-            <Frame className="frame" autoSize={false}>
-              <Track
-                currentView={selectedStore.id}
-                contain={false}
-                onViewChange={onViewChange}
-                className="track"
-              >
-                {stores.map((store) => (
-                  <View className="store-cell" key={store.id}>
-                    <StoreHeading
-                      store={store}
-                      selectedStore={selectedStore}
-                      onTapped={setSelectedStoreId}
-                      loadoutMenuRef={detachedLoadoutMenu}
-                    />
-                  </View>
-                ))}
-              </Track>
-            </Frame>
-          </ViewPager>
-        </ScrollClassDiv>
-
-        <div className="detached" ref={detachedLoadoutMenu} />
-
-        <Hammer direction="DIRECTION_HORIZONTAL" onSwipe={handleSwipe}>
-          <div>
-            <StoresInventory
-              stores={[selectedStore]}
-              selectedCategoryId={selectedCategoryId}
-              vault={vault}
-              currentStore={currentStore}
-              buckets={buckets}
-            />
-          </div>
-        </Hammer>
-
-        <div className="category-options">
-          {Object.keys(buckets.byCategory)
-            .filter((category) => category !== 'Postmaster')
-            .map((category) => (
-              <div
-                key={category}
-                onClick={() => handleCategoryChange(category)}
-                className={clsx({ selected: category === selectedCategoryId })}
-              >
-                {t(`Bucket.${category}`)}
-              </div>
-            ))}
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div
-      className={`inventory-content destiny${selectedStore.destinyVersion}`}
-      role="main"
-      aria-label={t('Header.Inventory')}
-    >
-      <ScrollClassDiv className="store-row store-header" scrollClass="sticky">
-        {stores.map((store, index) => (
-          <div
-            className={clsx('store-cell', { vault: store.isVault })}
-            key={store.id}
-            style={storeBackgroundColor(store, index)}
-          >
-            <StoreHeading store={store} />
-            <StoreStats store={store} />
-          </div>
-        ))}
-      </ScrollClassDiv>
-      <StoresInventory
-        stores={stores}
-        vault={vault}
-        currentStore={currentStore}
-        buckets={buckets}
-      />
-    </div>
+  return isPhonePortrait ? (
+    <PhoneStores stores={stores} buckets={buckets} />
+  ) : (
+    <DesktopStores stores={stores} buckets={buckets} />
   );
-}
-
-/** Is there any store that has an item in any of the buckets in this category? */
-function categoryHasItems(
-  allBuckets: InventoryBuckets,
-  category: string,
-  stores: DimStore[],
-  currentStore: DimStore
-): boolean {
-  const buckets = allBuckets.byCategory[category];
-  return buckets.some((bucket) => {
-    const storesToSearch = bucket.accountWide && !stores[0].isVault ? [currentStore] : stores;
-    return storesToSearch.some((s) => findItemsByBucket(s, bucket.hash).length > 0);
-  });
 }
 
 export default connect<StoreProps>(mapStateToProps)(Stores);
-
-interface InventoryContainerProps {
-  selectedCategoryId?: string;
-  buckets: InventoryBuckets;
-  stores: DimStore[];
-  currentStore: DimStore;
-  vault: DimStore;
-}
-
-function CollapsibleContainer({
-  buckets,
-  category,
-  stores,
-  currentStore,
-  inventoryBucket,
-  vault,
-  isPhonePortrait,
-}: {
-  category: string;
-  inventoryBucket: InventoryBucket[];
-  isPhonePortrait?: boolean;
-} & InventoryContainerProps) {
-  if (!categoryHasItems(buckets, category, stores, currentStore)) {
-    return null;
-  }
-
-  return (
-    <InventoryCollapsibleTitle title={t(`Bucket.${category}`)} sectionId={category} stores={stores}>
-      {/*
-          t('Bucket.Inventory')
-          t('Bucket.Postmaster')
-          t('Bucket.General')
-          t('Bucket.Progress')
-          t('Bucket.Unknown')
-        */}
-      {inventoryBucket.map((bucket) => (
-        <StoreBuckets
-          key={bucket.hash}
-          bucket={bucket}
-          stores={stores}
-          vault={vault}
-          currentStore={currentStore}
-          isPhonePortrait={isPhonePortrait}
-        />
-      ))}
-    </InventoryCollapsibleTitle>
-  );
-}
-
-function StoresInventory(props: InventoryContainerProps) {
-  const { selectedCategoryId, buckets, stores, currentStore, vault } = props;
-
-  const showPostmaster =
-    (currentStore.destinyVersion === 2 && selectedCategoryId === 'Inventory') ||
-    (currentStore.destinyVersion === 1 && selectedCategoryId === 'General');
-
-  if (selectedCategoryId) {
-    return (
-      <>
-        {selectedCategoryId === 'Armor' && (
-          <StoreStats
-            store={stores[0]}
-            style={{ ...storeBackgroundColor(stores[0], 0, true, true), paddingBottom: 8 }}
-          />
-        )}
-        {showPostmaster &&
-          buckets.byCategory['Postmaster'].map((bucket) => (
-            <StoreBuckets
-              key={bucket.hash}
-              bucket={bucket}
-              stores={stores}
-              vault={vault}
-              currentStore={currentStore}
-              labels={true}
-              isPhonePortrait={true}
-            />
-          ))}
-        {buckets.byCategory[selectedCategoryId].map((bucket) => (
-          <StoreBuckets
-            key={bucket.hash}
-            bucket={bucket}
-            stores={stores}
-            vault={vault}
-            currentStore={currentStore}
-            labels={true}
-            isPhonePortrait={true}
-          />
-        ))}
-      </>
-    );
-  }
-
-  return (
-    <>
-      {Object.entries(buckets.byCategory).map(([category, inventoryBucket]) => (
-        <CollapsibleContainer
-          key={category}
-          {...props}
-          category={category}
-          inventoryBucket={inventoryBucket}
-        />
-      ))}
-      {stores[0].destinyVersion === 1 && <D1ReputationSection stores={stores} />}
-    </>
-  );
-}
